refactor: split merger_ci.py into submodules
The 1146-line merger_ci.py is split into five focused modules:

- cmd_process.py  — flask merge / flask process (SBOM ingestion)
- cmd_export.py   — flask export / flask report (SBOM & report output)
- cmd_assessments.py — flask export/import-custom-assessments
- cmd_scans.py    — flask list-projects / list-scans / delete-scan
- cmd_vuln_scan.py — flask nvd-scan / flask osv-scan

merger_ci.py is now a thin orchestrator (~70 lines) that imports from\nthese modules, registers all CLI commands via init_app(), and\nre-exports the public symbols (post_treatment, etc.) that webapp.py depends on. All 1358 tests pass.
